La Simulación de Pagos: Implementación de Modelos y Repositorio

+++ src/main/java/com/DTOs/PaymentRequest.java
@@ -13,7 +13,7 @@ public class PaymentRequest {
     private Long idPedido;
     
     @Nonnull()
-    private Integer idMetodoPago;
+    private Long idMetodoPago;
     
     @Nonnull()
     private BigDecimal monto;
